requirements.txt
# Web Framework & API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx>=0.24.0
requests==2.31.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
supabase==2.3.0
alembic==1.13.1

# Environment & Configuration
python-dotenv==1.0.0

# AI & Chat (for future chat interface)
anthropic==0.25.0

# Notifications
slack-sdk==3.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Scheduling (for automated syncs)
APScheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
loguru==0.7.2
sentry-sdk[fastapi]==1.39.2
